+++ app/src/main/java/com/android/shaftschematic/ui/screen/ShaftScreen.kt
@@ -129,11 +129,8 @@ import com.android.shaftschematic.ui.util.buildBodyTitleById
 import com.android.shaftschematic.ui.util.buildLinerTitleById
 import com.android.shaftschematic.ui.util.buildTaperTitleById
 import com.android.shaftschematic.ui.util.buildThreadTitleById
-import com.android.shaftschematic.ui.util.bodyWarningMessage
 import com.android.shaftschematic.ui.util.freeToEndSignedMm
-import com.android.shaftschematic.ui.util.linerWarningMessage
 import com.android.shaftschematic.ui.util.startOverlapErrorMm
-import com.android.shaftschematic.ui.util.taperWarningMessage
 import com.android.shaftschematic.ui.util.threadWarningMessage
 import com.android.shaftschematic.ui.viewmodel.SnapConfig
 import com.android.shaftschematic.ui.viewmodel.SessionAddDefaults